initializer tests
