+++ src/codegen/tests/initialization_test.rs
@@ -1,4 +1,4 @@
 mod global_initializers;
-mod init_functions;
+mod complex_initializers;
 mod pou_initializers;
 mod type_initializers;
